Renamed controller folder to avoid merge conflict
