+++ src/main/java/org/launchcode/sprinklespre/controller/ApiCourseCountroller.java
@@ -1,12 +1,4 @@
-package org.launchcode.sprinklespre.controllers.api;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Optional;
+package org.launchcode.sprinklespre.controller;
 
 //TODO: Set up @CrossOrigin to localhost path for React
 //TODO: @RestController
