+++ src/main/java/org/launchcode/sprinklespre/controller/UserController.java
@@ -1,0 +1,5 @@
+package org.launchcode.sprinklespre.controller;
+
+public class UserController {
+
+}
